requirements.txt
pycrypto==2.6.1
pycryptodome==3.18.0
tqdm==4.65.0
